requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyMySQL==1.1.0
line-bot-sdk==3.5.1
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug>=3.0.0